@@ -49,12 +49,9 @@ const config = defineConfig([
   {
     input: 'src/htmlminifier.js',
     output: {
-      file: 'dist/htmlminifier.mjs',
-      format: 'es'
+      file: 'dist/htmlminifier.cjs',
+      format: 'cjs'
     },
-    plugins: [
-      commonjs()
-    ],
     external: ['clean-css', 'terser', 'entities', 'relateurl']
   }
 ]);
