@@ -1,10 +1,10 @@
-'use strict';
-
-function createMap(values, ignoreCase) {
+export function createMap(values, ignoreCase) {
   const map = {};
+
   values.forEach(function (value) {
     map[value] = 1;
   });
+
   return ignoreCase
     ? function (value) {
       return map[value.toLowerCase()] === 1;
@@ -14,19 +14,18 @@ function createMap(values, ignoreCase) {
     };
 }
 
-async function replaceAsync(str, regex, asyncFn) {
+export async function replaceAsync(str, regex, asyncFn) {
   const promises = [];
+
   str.replace(regex, (match, ...args) => {
     const promise = asyncFn(match, ...args);
     promises.push(promise);
   });
+
   const data = await Promise.all(promises);
   return str.replace(regex, () => data.shift());
 }
 
-exports.createMap = createMap;
-exports.createMapFromString = function (values, ignoreCase) {
+export function createMapFromString(values, ignoreCase) {
   return createMap(values.split(/,/), ignoreCase);
-};
-
-exports.replaceAsync = replaceAsync;
+}
